+++ near-sdk/src/utils/storage_key_impl.rs
@@ -1,5 +1,3 @@
-use borsh::BorshSerialize;
-
 /// Converts Self into a [`Vec<u8>`] that is used for a storage key through [`into_storage_key`].
 ///
 /// [`into_storage_key`]: IntoStorageKey::into_storage_key
@@ -33,34 +31,5 @@ impl IntoStorageKey for u8 {
     #[inline]
     fn into_storage_key(self) -> Vec<u8> {
         vec![self]
-    }
-}
-
-/// Converts a Borsh serializable object into a `Vec<u8>` that is used for a storage key.
-///
-/// ```
-/// use near_sdk::borsh::BorshSerialize;
-/// use near_sdk::BorshIntoStorageKey;
-/// use near_sdk::collections::LookupMap;
-///
-/// #[derive(BorshSerialize)]
-///  enum StorageKey {
-///     FungibleToken,
-///     Metadata { sub_key: String },
-/// }
-///
-/// impl BorshIntoStorageKey for StorageKey {}
-///
-/// let lookup_map_1: LookupMap<u64, String> = LookupMap::new(StorageKey::Metadata { sub_key: String::from("yo") });
-/// let lookup_map_2: LookupMap<String, String> = LookupMap::new(StorageKey::FungibleToken);
-/// ```
-pub trait BorshIntoStorageKey: BorshSerialize {}
-
-impl<T> IntoStorageKey for T
-where
-    T: BorshIntoStorageKey,
-{
-    fn into_storage_key(self) -> Vec<u8> {
-        self.try_to_vec().unwrap()
     }
 }
